Backend/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.3
django-braces==1.14.0
django-cors-headers==3.2.1
django-heroku==0.3.1
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.8
oauthlib==3.1.0
psycopg2==2.8.4
PyJWT==1.7.1
python-dotenv==0.11.0
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
lml==0.0.9
pyexcel==0.5.15
pyexcel-io==0.5.20
texttable==1.6.2
xlrd==1.2.0
et-xmlfile==1.0.1
jdcal==1.4.1
openpyxl==3.0.3
pyexcel-xlsx==0.5.8
pyfcm==1.4.7
httplib2==0.9.2
twilio==5.6.0
wheel==0.24.0